requirements.txt
pyspark==3.5.1
pandas>=2.2.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pyyaml>=6.0.1
